+++ src/sdk/PnP.Core/Model/Base/BatchResult.cs
@@ -41,12 +41,12 @@ namespace PnP.Core.Model
         {
             get
             {
-                if (!IsAvailable)
+                if (!IsAvailable && BatchRequest.ExecutionNeeded)
                 {
                     throw new InvalidOperationException(PnPCoreResources.Exception_BatchResult_BatchNotYetExecuted);
                 }
 
-                if (BatchRequest.ApiCall.RawRequest)
+                if (BatchRequest.ApiCall.RawRequest || BatchRequest.ApiCall.RawResultsHandler != null)
                 {
                     return result;
                 }